14_Documentation/my_proj/requirements.txt
alabaster==0.7.12
asgiref==3.3.4
Babel==2.9.1
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.4
django-filter==2.4.0
djangorestframework==3.12.4
docutils==0.17.1
drf-yasg==1.20.0
idna==2.10
imagesize==1.2.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
packaging==20.9
Pygments==2.9.0
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.17.9
ruamel.yaml.clib==0.2.2
snowballstemmer==2.1.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.5
